[FLINK-3229] Move all manual Kinesis connector tests to org.apache.flink.streaming.connectors.kinesis.manualtests
