+++ flink-streaming-connectors/flink-connector-kinesis/src/test/java/org/apache/flink/streaming/connectors/kinesis/manualtests/ManualProducerTest.java
@@ -14,20 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.streaming.connectors.kinesis;
+package org.apache.flink.streaming.connectors.kinesis.manualtests;
 
-import org.apache.flink.api.common.functions.FlatMapFunction;
 import org.apache.flink.api.java.utils.ParameterTool;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.streaming.connectors.kinesis.config.KinesisConfigConstants;
+import org.apache.flink.streaming.connectors.kinesis.FlinkKinesisProducer;
+import org.apache.flink.streaming.connectors.kinesis.KinesisPartitioner;
 import org.apache.flink.streaming.connectors.kinesis.examples.ProduceIntoKinesis;
 import org.apache.flink.streaming.connectors.kinesis.serialization.KinesisSerializationSchema;
-import org.apache.flink.streaming.util.serialization.SimpleStringSchema;
-import org.apache.flink.util.Collector;
 
 import java.nio.ByteBuffer;
-import java.util.Properties;
 
 /**
  * This is a manual test for the AWS Kinesis connector in Flink.
@@ -36,10 +33,12 @@ import java.util.Properties;
  *  - A custom KinesisSerializationSchema
  *  - A custom KinesisPartitioner
  *
+ *  The streams "test-flink" and "flink-test-2" must exist.
+ *
  * Invocation:
  * --region eu-central-1 --accessKey XXXXXXXXXXXX --secretKey XXXXXXXXXXXXXXXX
  */
-public class ManualConsumerProducerTest {
+public class ManualProducerTest {
 
 	public static void main(String[] args) throws Exception {
 		ParameterTool pt = ParameterTool.fromArgs(args);
@@ -79,33 +78,6 @@ public class ManualConsumerProducerTest {
 			}
 		});
 		simpleStringStream.addSink(kinesis);
-
-
-		// consuming topology
-		Properties consumerProps = new Properties();
-		consumerProps.setProperty(KinesisConfigConstants.CONFIG_AWS_CREDENTIALS_PROVIDER_BASIC_ACCESSKEYID, pt.getRequired("accessKey"));
-		consumerProps.setProperty(KinesisConfigConstants.CONFIG_AWS_CREDENTIALS_PROVIDER_BASIC_SECRETKEY, pt.getRequired("secretKey"));
-		consumerProps.setProperty(KinesisConfigConstants.CONFIG_AWS_REGION, pt.getRequired("region"));
-		DataStream<String> consuming = see.addSource(new FlinkKinesisConsumer<>("test-flink", new SimpleStringSchema(), consumerProps));
-		// validate consumed records for correctness
-		consuming.flatMap(new FlatMapFunction<String, String>() {
-			@Override
-			public void flatMap(String value, Collector<String> out) throws Exception {
-				String[] parts = value.split("-");
-				try {
-					long l = Long.parseLong(parts[0]);
-					if(l < 0) {
-						throw new RuntimeException("Negative");
-					}
-				} catch(NumberFormatException nfe) {
-					throw new RuntimeException("First part of '" + value + "' is not a valid numeric type");
-				}
-				if(parts[1].length() != 12) {
-					throw new RuntimeException("Second part of '" + value + "' doesn't have 12 characters");
-				}
-			}
-		});
-		consuming.print();
 
 		see.execute();
 	}
